Make NRD::Packet have a limit of data size, and handle the new condition in the daemon.
This is to prevent a malicious / unwanted client from connecting to the socket, and consuming all the memory on the system.

Test suite injects garbage into the server. The imporant thing is that the server doesn't die when strange things arrive to it (the server can still continue recieving results).